Can handle clients and a bike now.

+++ BikeLib/Bike.cs
@@ -7,7 +7,7 @@ namespace BikeLib
     public class Bike
     {
         public int Id { get; set; }
-        public int Price { get; set; }
+        public double Price { get; set; }
         public int Gear { get; set; }
         public string Color { get; set; }
 
@@ -16,7 +16,7 @@ namespace BikeLib
 
         }
 
-        public Bike(int id, int price, int gear, string color)
+        public Bike(int id, double price, int gear, string color)
         {
             Id = id;
             Price = price;
